Move ensure_open to core serialization
